+++ server/tests/security/test_oauth_callback_redaction.py
@@ -20,11 +20,6 @@ _server_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
 if os.path.abspath(_server_dir) not in sys.path:
     sys.path.insert(0, os.path.abspath(_server_dir))
 
-# REDIS_URL is required at import time by utils.auth.oauth2_state_cache.
-# Set a test default before _build_app imports that module so the collection
-# step does not raise even when Redis is not running.
-os.environ.setdefault("REDIS_URL", "redis://localhost:6379/0")
-
 
 _CLIENT_SECRET_MARKER = "client_secret_DO_NOT_ECHO_XYZ"
 
@@ -36,74 +31,9 @@ def _exc_carrying_client_secret() -> Exception:
     )
 
 
-def _evict_route_modules(*prefixes: str) -> None:
-    to_delete = [mod for mod in sys.modules if any(mod == p or mod.startswith(p + ".") for p in prefixes)]
-    for mod in to_delete:
-        del sys.modules[mod]
-
-
-def _install_audit_stub() -> MagicMock:
-    stub = MagicMock(name="routes.audit_routes")
-    stub.record_audit_event = MagicMock()
-    sys.modules["routes.audit_routes"] = stub
-    return stub
-
-
-def _make_test_app():
-    """Return a minimal Flask app for OAuth callback testing.
-
-    TESTING=True enables exception propagation for cleaner assertions.
-    Aurora has no Flask-side CSRF middleware, so this does not disable any
-    CSRF protection — the trust boundary is the Next.js proxy layer.
-    """
-    from flask import Flask as _Flask
-    tmp_app = _Flask(__name__)
-    tmp_app.config["TESTING"] = True  # NOSONAR
-    return tmp_app
-
-
-def _stub_limiter() -> None:
-    """Disable rate-limiting in the already-registered limiter stub."""
-    limiter_stub = sys.modules["utils.web.limiter_ext"]
-    limiter_stub.limiter = MagicMock()
-    limiter_stub.limiter.limit = lambda *a, **kw: (lambda f: f)
-
-
-def _patch_rbac_deps(monkeypatch, oauth_state_endpoint: str) -> None:
-    """Monkeypatch RBAC, state-cache, token-management, and DB pool deps."""
-    import utils.auth.rbac_decorators as rbac_mod
-    import utils.auth.oauth2_state_cache as state_cache_mod
-    import utils.auth.token_management as tok_mod
-    import utils.db.connection_pool as pool_mod
-
-    monkeypatch.setattr(rbac_mod, "get_user_id_from_request", MagicMock(return_value="u-1"))
-    monkeypatch.setattr(rbac_mod, "get_org_id_from_request", MagicMock(return_value="org-1"))
-    monkeypatch.setattr(rbac_mod, "enforce_with_reload", MagicMock(return_value=True))
-    monkeypatch.setattr(rbac_mod, "_audit_auth_failure", MagicMock())
-    monkeypatch.setattr(
-        state_cache_mod,
-        "retrieve_oauth2_state",
-        MagicMock(return_value={"user_id": "u-1", "endpoint": oauth_state_endpoint}),
-    )
-    monkeypatch.setattr(tok_mod, "store_tokens_in_db", MagicMock())
-    monkeypatch.setattr(pool_mod, "db_pool", MagicMock())
-
-
 # ---------------------------------------------------------------------------
 # Atlassian callback
 # ---------------------------------------------------------------------------
-
-
-_ATLASSIAN_HEAVY = (
-    "connectors.atlassian_auth",
-    "connectors.atlassian_auth.auth",
-    "connectors.confluence_connector",
-    "connectors.confluence_connector.client",
-    "connectors.jira_connector",
-    "connectors.jira_connector.client",
-    "utils.web.limiter_ext",
-    "config.rate_limiting",
-)
 
 
 class TestAtlassianOAuthCallbackRedaction:
@@ -112,31 +42,14 @@ class TestAtlassianOAuthCallbackRedaction:
     """
 
     @pytest.fixture(autouse=True)
-    def _setup(self, monkeypatch):
-        for pkg in _ATLASSIAN_HEAVY:
-            sys.modules.setdefault(pkg, MagicMock())
-        _install_audit_stub()
-        _stub_limiter()
-
-    def _build_app(self, monkeypatch):
-        _evict_route_modules("routes.atlassian")
-        _patch_rbac_deps(monkeypatch, "atlassian:confluence")
+    def _inject_failure(self, atlassian_oauth_app):
         sys.modules["connectors.atlassian_auth.auth"].exchange_code_for_token = (
             MagicMock(side_effect=_exc_carrying_client_secret())
         )
+        self.app = atlassian_oauth_app
 
-        from routes.atlassian.atlassian_routes import atlassian_bp
-
-        app = _make_test_app()
-        app.register_blueprint(atlassian_bp, url_prefix="/atlassian")
-        return app
-
-    def test_exchange_failure_response_does_not_contain_secret(
-        self, monkeypatch, caplog
-    ):
-        app = self._build_app(monkeypatch)
-
-        with app.test_client() as client:
+    def test_exchange_failure_response_does_not_contain_secret(self, caplog):
+        with self.app.test_client() as client:
             with caplog.at_level(logging.DEBUG):
                 resp = client.post(
                     "/atlassian/connect",
@@ -155,10 +68,8 @@ class TestAtlassianOAuthCallbackRedaction:
                 f"client_secret found in log record: {record.getMessage()!r}"
             )
 
-    def test_exchange_failure_returns_generic_error(self, monkeypatch):
-        app = self._build_app(monkeypatch)
-
-        with app.test_client() as client:
+    def test_exchange_failure_returns_generic_error(self):
+        with self.app.test_client() as client:
             resp = client.post(
                 "/atlassian/connect",
                 json={
@@ -180,46 +91,20 @@ class TestAtlassianOAuthCallbackRedaction:
 # ---------------------------------------------------------------------------
 
 
-_NOTION_HEAVY = (
-    "connectors.notion_connector",
-    "connectors.notion_connector.auth",
-    "connectors.notion_connector.client",
-    "utils.web.limiter_ext",
-    "config.rate_limiting",
-)
-
-
 class TestNotionOAuthCallbackRedaction:
     """The Notion connect route must not echo ``client_secret`` material
     when a token exchange fails due to a network-level exception.
     """
 
     @pytest.fixture(autouse=True)
-    def _setup(self, monkeypatch):
-        for pkg in _NOTION_HEAVY:
-            sys.modules.setdefault(pkg, MagicMock())
-        _install_audit_stub()
-        _stub_limiter()
-
-    def _build_app(self, monkeypatch):
-        _evict_route_modules("routes.notion")
-        _patch_rbac_deps(monkeypatch, "notion")
+    def _inject_failure(self, notion_oauth_app):
         sys.modules["connectors.notion_connector.auth"].exchange_code_for_token = (
             MagicMock(side_effect=_exc_carrying_client_secret())
         )
+        self.app = notion_oauth_app
 
-        from routes.notion.notion_routes import notion_bp
-
-        app = _make_test_app()
-        app.register_blueprint(notion_bp, url_prefix="/notion")
-        return app
-
-    def test_exchange_failure_response_does_not_contain_secret(
-        self, monkeypatch, caplog
-    ):
-        app = self._build_app(monkeypatch)
-
-        with app.test_client() as client:
+    def test_exchange_failure_response_does_not_contain_secret(self, caplog):
+        with self.app.test_client() as client:
             with caplog.at_level(logging.DEBUG):
                 resp = client.post(
                     "/notion/oauth/callback",
@@ -233,10 +118,8 @@ class TestNotionOAuthCallbackRedaction:
                 f"client_secret found in log record: {record.getMessage()!r}"
             )
 
-    def test_exchange_failure_returns_generic_error(self, monkeypatch):
-        app = self._build_app(monkeypatch)
-
-        with app.test_client() as client:
+    def test_exchange_failure_returns_generic_error(self):
+        with self.app.test_client() as client:
             resp = client.post(
                 "/notion/oauth/callback",
                 json={"code": "auth-code-xyz", "state": "state-token-abc"},
